Created new balance retrival, on new endpoints. Add the unit tests and code adjustments.

+++ MultiAcctAPI/Services/AccountService.cs
@@ -1,6 +1,8 @@
+using Microsoft.EntityFrameworkCore;
 using MultiAcctAPI.Data;
+using MultiAcctAPI.Interfaces;
 using MultiAcctAPI.Models;
-using MultiAcctAPI.Services.Interfaces;
+using MultiAcctAPI.ModelsAuxiliary;
 
 namespace MultiAcctAPI.Services
 {
@@ -13,9 +15,9 @@ namespace MultiAcctAPI.Services
             _context = context;
         }
 
-        public IEnumerable<Account> GetAccountsByUserId(Guid userId)
+        public async Task<IEnumerable<Account>> GetAccountsByUserIdAsync(Guid userId)
         {
-            return _context.Accounts
+            return await _context.Accounts
                 .Where(a => a.UserId == userId)
                 .Select(a => new Account
                 {
@@ -25,41 +27,71 @@ namespace MultiAcctAPI.Services
                     CurrentBalance = a.CurrentBalance,
                     UserId = a.UserId
                 })
-                .ToList();
+                .ToListAsync();
         }
 
-        public Account GetAccountById(Guid accountId)
+        public async Task<Account?> GetAccountByIdAsync(Guid accountId)
         {
-            return _context.Accounts.SingleOrDefault(a => a.AccountId == accountId);
-        }
-
-        public Account CreateAccount(Account account)
-        {
-            account.AccountId = Guid.NewGuid();
-            _context.Accounts.Add(account);
-            _context.SaveChanges();
+            var account = await _context.Accounts.FirstOrDefaultAsync(a => a.AccountId == accountId);
+            if (account == null)
+            {
+                throw new KeyNotFoundException($"Account with ID {accountId} not found.");
+            }
             return account;
         }
 
-        public void UpdateAccount(Account account)
+        public async Task<Account> CreateAccountAsync(Account account)
         {
-            var existingAccount = GetAccountById(account.AccountId);
+            account.AccountId = Guid.NewGuid();
+            await _context.Accounts.AddAsync(account);
+            await _context.SaveChangesAsync();
+            return account;
+        }
+
+        public async Task UpdateAccountAsync(Account account)
+        {
+            var existingAccount = await GetAccountByIdAsync(account.AccountId);
             if (existingAccount != null)
             {
                 existingAccount.AccountName = account.AccountName;
                 existingAccount.CurrentBalance = account.CurrentBalance;
-                _context.SaveChanges();
+                await _context.SaveChangesAsync();
             }
         }
 
-        public void DeleteAccount(Guid accountId)
+        public async Task DeleteAccountAsync(Guid accountId)
         {
-            var account = GetAccountById(accountId);
+            var account = await GetAccountByIdAsync(accountId);
             if (account != null)
             {
+                if (account.CurrentBalance != 0)
+                {
+                    throw new InvalidOperationException("Account current balance should be 0 (zero).");
+                }
                 _context.Accounts.Remove(account);
-                _context.SaveChanges();
+                await _context.SaveChangesAsync();
             }
+        }
+
+        public async Task<decimal> GetAccountBalanceAsync(Guid accountId)
+        {
+            var account = await _context.Accounts.FirstOrDefaultAsync(a => a.AccountId == accountId);
+            if (account == null)
+                throw new InvalidOperationException("Account not found.");
+
+            return account.CurrentBalance;
+        }
+
+        public async Task<IEnumerable<AccountSummary>> GetUserAccountSummariesAsync(Guid userId)
+        {
+            return await _context.Accounts
+                .Where(a => a.UserId == userId)
+                .Select(a => new AccountSummary
+                {
+                    AccountId = a.AccountId,
+                    Balance = a.CurrentBalance
+                })
+                .ToListAsync();
         }
     }
 }
